pyproject.toml
[tool.poetry]
name = "unu-api"
version = "0.1.0"
description = "Full REST API for unu. A minimalistic event manager."
authors = ["emanuelosva <emanuelosva@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
passlib = "^1.7.2"
pydantic = "^1.6.1"
python-jose = "^3.2.0"
python-multipart = "^0.0.5"
redis = "^3.5.3"
rq = "^1.5.2"
sendgrid = "^6.4.7"
uvicorn = "^0.12.1"
six = "^1.15.0"
python-dotenv = "^0.14.0"
bcrypt = "^3.2.0"
dnspython = "^2.0.0"
google-cloud-storage = "^1.31.2"
tortoise-orm = "^0.16.16"
asyncpg = "^0.21.0"
aerich = "^0.2.5"

[tool.poetry.dev-dependencies]
pytest = "^6.1.0"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.10.1"
pytest-xdist = "^2.1.0"
flake8 = "^3.8.3"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
